pyproject.toml
[project]
name = "we0-index"
version = "0.1.0"
description = "We0 Index"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "chromadb>=0.6.3",
    "click>=8.0.0",
    "fastapi>=0.115.7",
    "gitpython>=3.1.42",
    "greenlet>=3.2.2",
    "httpx>=0.24.0",
    "loguru>=0.7.3",
    "mcp[cli]>=1.9.2",
    "numpy>=1.24.0",
    "openai",
    "psycopg[binary,pool]>=3.2.4",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.13.2",
    "sqlalchemy>=2.0.41",
    "tiktoken>=0.8.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-css>=0.23.2",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-typescript>=0.23.2",
    "uvicorn>=0.34.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
